requirements.txt
pymongo[srv]
certifi
from_root
numpy
pandas
scikit-learn
PyYAML
scipy

-e . 